Update GHA CI/CD checkout for version string (#173)
Currently, the locally built copies of the liboscal-java library does
correctly encode version information based on short ID for commits and
tags. This behavior functions correctly because the `git clone` for
normal operations retrieves full history so you can properly analyze
commits and find the latest tag.

This behavior does not work with the default behavior (and arguments),
as of actions/checkout@c85c95e3d7251135ab7dc9ce3241c5835cc595a9, does
not get that history. This commit configures GitHub CI/CD with this
action to fetch the full history of the repo to do it correctly.